+++ src/library/checkpoint/Checkpoint.h
@@ -28,12 +28,12 @@ namespace libtas {
 namespace Checkpoint
 {
     void setSavestatePath(std::string path);
-    void setParentSavestatePath(std::string path);
     void setBaseSavestatePath(std::string path);
 
     void setSavestateIndex(int index);
     void setBaseSavestateIndex(int index);
-    void setParentSavestateIndex(int index);
+
+    void setCurrentToParent();
 
     bool checkCheckpoint();
     bool checkRestore();