+++ src/program/KeyMapping.h
@@ -65,6 +65,7 @@ typedef int HotKeyType; enum
     HOTKEY_LOADSTATE7,
     HOTKEY_LOADSTATE8,
     HOTKEY_LOADSTATE9,
+    HOTKEY_LOADSTATE_BACKTRACK,
     HOTKEY_TOGGLE_ENCODE, // Start/stop audio/video encoding
     HOTKEY_CALIBRATE_MOUSE, // Calibrate mouse cursor position
     HOTKEY_LEN